Extract TOML parsing helpers
